requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
firebase-admin==6.0.1
requests==2.31.0
